+++ opentelemetry-sdk/src/logs/mod.rs
@@ -87,7 +87,7 @@ mod tests {
         let log = exported_logs
             .first()
             .expect("Atleast one log is expected to be present.");
-        assert_eq!(log.instrumentation.name, "test-logger");
+        assert_eq!(log.instrumentation.name(), "test-logger");
         assert_eq!(log.record.severity_number, Some(Severity::Error));
         assert_eq!(log.record.attributes_len(), 10);
         for i in 1..=10 {
@@ -111,14 +111,13 @@ mod tests {
 
         let logger = provider.logger_with_scope(scope);
         let instrumentation_scope = logger.instrumentation_scope();
-        let attributes = &instrumentation_scope.attributes;
-        assert_eq!(instrumentation_scope.name, "test_logger");
+        assert_eq!(instrumentation_scope.name(), "test_logger");
         assert_eq!(
-            instrumentation_scope.schema_url,
-            Some("https://opentelemetry.io/schema/1.0.0".into())
+            instrumentation_scope.schema_url(),
+            Some("https://opentelemetry.io/schema/1.0.0")
         );
-        assert_eq!(attributes.len(), 1);
-        assert_eq!(attributes[0].key, "test_k".into());
-        assert_eq!(attributes[0].value, "test_v".into());
+        assert!(instrumentation_scope
+            .attributes()
+            .eq(&[KeyValue::new("test_k", "test_v")]));
     }
 }