+++ opentelemetry-sdk/src/metrics/meter.rs
@@ -76,7 +76,7 @@ impl SdkMeter {
         if let Err(err) = validation_result {
             otel_error!(
                 name: "InstrumentCreationFailed",
-                meter_name = self.scope.name.as_ref(),
+                meter_name = self.scope.name(),
                 instrument_name = builder.name.as_ref(),
                 message = "Measurements from this Counter will be ignored.",
                 reason = format!("{}", err)
@@ -98,7 +98,7 @@ impl SdkMeter {
             Err(err) => {
                 otel_error!(
                     name: "InstrumentCreationFailed",
-                    meter_name = self.scope.name.as_ref(),
+                    meter_name = self.scope.name(),
                     instrument_name = builder.name.as_ref(),
                     message = "Measurements from this Counter will be ignored.",
                     reason = format!("{}", err)
@@ -120,7 +120,7 @@ impl SdkMeter {
         if let Err(err) = validation_result {
             otel_error!(
                 name: "InstrumentCreationFailed", 
-                meter_name = self.scope.name.as_ref(),
+                meter_name = self.scope.name(),
                 instrument_name = builder.name.as_ref(),
                 message = "Callbacks for this ObservableCounter will not be invoked.",
                 reason = format!("{}", err));
@@ -138,7 +138,7 @@ impl SdkMeter {
                 if ms.is_empty() {
                     otel_error!(
                         name: "InstrumentCreationFailed",
-                        meter_name = self.scope.name.as_ref(),
+                        meter_name = self.scope.name(),
                         instrument_name = builder.name.as_ref(),
                         message = "Callbacks for this ObservableCounter will not be invoked. Check View Configuration."
                     );
@@ -158,7 +158,7 @@ impl SdkMeter {
             Err(err) => {
                 otel_error!(
                     name: "InstrumentCreationFailed",
-                    meter_name = self.scope.name.as_ref(),
+                    meter_name = self.scope.name(),
                     instrument_name = builder.name.as_ref(),
                     message = "Callbacks for this ObservableCounter will not be invoked.",
                     reason = format!("{}", err));
@@ -179,7 +179,7 @@ impl SdkMeter {
         if let Err(err) = validation_result {
             otel_error!(
                 name: "InstrumentCreationFailed", 
-                meter_name = self.scope.name.as_ref(),
+                meter_name = self.scope.name(),
                 instrument_name = builder.name.as_ref(),
                 message = "Callbacks for this ObservableUpDownCounter will not be invoked.",
                 reason = format!("{}", err));
@@ -197,7 +197,7 @@ impl SdkMeter {
                 if ms.is_empty() {
                     otel_error!(
                         name: "InstrumentCreationFailed",
-                        meter_name = self.scope.name.as_ref(),
+                        meter_name = self.scope.name(),
                         instrument_name = builder.name.as_ref(),
                         message = "Callbacks for this ObservableUpDownCounter will not be invoked. Check View Configuration."
                     );
@@ -217,7 +217,7 @@ impl SdkMeter {
             Err(err) => {
                 otel_error!(
                     name: "InstrumentCreationFailed",
-                    meter_name = self.scope.name.as_ref(),
+                    meter_name = self.scope.name(),
                     instrument_name = builder.name.as_ref(),
                     message = "Callbacks for this ObservableUpDownCounter will not be invoked.",
                     reason = format!("{}", err));
@@ -238,7 +238,7 @@ impl SdkMeter {
         if let Err(err) = validation_result {
             otel_error!(
                 name: "InstrumentCreationFailed", 
-                meter_name = self.scope.name.as_ref(),
+                meter_name = self.scope.name(),
                 instrument_name = builder.name.as_ref(),
                 message = "Callbacks for this ObservableGauge will not be invoked.",
                 reason = format!("{}", err));
@@ -256,7 +256,7 @@ impl SdkMeter {
                 if ms.is_empty() {
                     otel_error!(
                         name: "InstrumentCreationFailed",
-                        meter_name = self.scope.name.as_ref(),
+                        meter_name = self.scope.name(),
                         instrument_name = builder.name.as_ref(),
                         message = "Callbacks for this ObservableGauge will not be invoked. Check View Configuration."
                     );
@@ -276,7 +276,7 @@ impl SdkMeter {
             Err(err) => {
                 otel_error!(
                     name: "InstrumentCreationFailed",
-                    meter_name = self.scope.name.as_ref(),
+                    meter_name = self.scope.name(),
                     instrument_name = builder.name.as_ref(),
                     message = "Callbacks for this ObservableGauge will not be invoked.",
                     reason = format!("{}", err));
@@ -297,7 +297,7 @@ impl SdkMeter {
         if let Err(err) = validation_result {
             otel_error!(
                 name: "InstrumentCreationFailed",
-                meter_name = self.scope.name.as_ref(),
+                meter_name = self.scope.name(),
                 instrument_name = builder.name.as_ref(),
                 message = "Measurements from this UpDownCounter will be ignored.",
                 reason = format!("{}", err)
@@ -319,7 +319,7 @@ impl SdkMeter {
             Err(err) => {
                 otel_error!(
                     name: "InstrumentCreationFailed",
-                    meter_name = self.scope.name.as_ref(),
+                    meter_name = self.scope.name(),
                     instrument_name = builder.name.as_ref(),
                     message = "Measurements from this UpDownCounter will be ignored.",
                     reason = format!("{}", err)
@@ -341,7 +341,7 @@ impl SdkMeter {
         if let Err(err) = validation_result {
             otel_error!(
                 name: "InstrumentCreationFailed",
-                meter_name = self.scope.name.as_ref(),
+                meter_name = self.scope.name(),
                 instrument_name = builder.name.as_ref(),
                 message = "Measurements from this Gauge will be ignored.",
                 reason = format!("{}", err)
@@ -363,7 +363,7 @@ impl SdkMeter {
             Err(err) => {
                 otel_error!(
                     name: "InstrumentCreationFailed",
-                    meter_name = self.scope.name.as_ref(),
+                    meter_name = self.scope.name(),
                     instrument_name = builder.name.as_ref(),
                     message = "Measurements from this Gauge will be ignored.",
                     reason = format!("{}", err)
@@ -385,7 +385,7 @@ impl SdkMeter {
         if let Err(err) = validation_result {
             otel_error!(
                 name: "InstrumentCreationFailed",
-                meter_name = self.scope.name.as_ref(),
+                meter_name = self.scope.name(),
                 instrument_name = builder.name.as_ref(),
                 message = "Measurements from this Histogram will be ignored.",
                 reason = format!("{}", err)
@@ -407,7 +407,7 @@ impl SdkMeter {
             Err(err) => {
                 otel_error!(
                     name: "InstrumentCreationFailed",
-                    meter_name = self.scope.name.as_ref(),
+                    meter_name = self.scope.name(),
                     instrument_name = builder.name.as_ref(),
                     message = "Measurements from this Histogram will be ignored.",
                     reason = format!("{}", err)