+++ opentelemetry-sdk/src/metrics/meter_provider.rs
@@ -465,18 +465,16 @@ mod tests {
         assert_eq!(provider.inner.meters.lock().unwrap().len(), 2);
 
         // these are different meters because meter names are case sensitive
-        let mut library = InstrumentationScope::builder("ABC")
-            .with_version("1.0.0")
-            .with_schema_url("http://example.com")
-            .build();
+        let make_scope = |name| {
+            InstrumentationScope::builder(name)
+                .with_version("1.0.0")
+                .with_schema_url("http://example.com")
+                .build()
+        };
 
-        let _meter6 = provider.meter_with_scope(library.clone());
-
-        library.name = "Abc".into();
-        let _meter7 = provider.meter_with_scope(library.clone());
-
-        library.name = "abc".into();
-        let _meter8 = provider.meter_with_scope(library);
+        let _meter6 = provider.meter_with_scope(make_scope("ABC"));
+        let _meter7 = provider.meter_with_scope(make_scope("Abc"));
+        let _meter8 = provider.meter_with_scope(make_scope("abc"));
 
         assert_eq!(provider.inner.meters.lock().unwrap().len(), 5);
     }